refactor: move YouTube ID parsing to library
