+++ src/components/recorder/youtube-reference.tsx
@@ -8,7 +8,11 @@ import {
   RecorderRuntime,
   type ReferenceVideoState,
 } from "../../lib/recorder/runtime";
-import { loadYouTubeApi, type YouTubePlayerApi } from "../../lib/youtube";
+import {
+  loadYouTubeApi,
+  parseYouTubeVideoId,
+  type YouTubePlayerApi,
+} from "../../lib/youtube";
 import { Button } from "../ui/button";
 import { FloatingPanel } from "../ui/floating-panel";
 
@@ -330,26 +334,4 @@ function YouTubeReference({
       )}
     </div>
   );
-}
-
-function parseYouTubeVideoId(value: string): string | undefined {
-  const input = value.trim();
-  if (/^[\w-]{11}$/.test(input)) {
-    return input;
-  }
-  try {
-    const url = new URL(input);
-    const host = url.hostname.replace(/^www\./, "");
-    let videoId: string | undefined;
-    if (host === "youtu.be") {
-      videoId = url.pathname.split("/")[1];
-    } else if (host === "youtube.com" || host === "m.youtube.com") {
-      videoId =
-        url.searchParams.get("v") ??
-        url.pathname.match(/^\/(?:embed|shorts)\/([^/]+)/)?.[1];
-    }
-    return videoId && /^[\w-]{11}$/.test(videoId) ? videoId : undefined;
-  } catch {
-    return undefined;
-  }
 }
